% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_token.R
\name{generate_token}
\alias{generate_token}
\title{Generate an API token}
\usage{
generate_token(username, password)
}
\arguments{
\item{username}{Impect login email}

\item{password}{Impect password}
}
\description{
generate_token generates a token that can be used to get data from the Impect API
}
