% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_competitions.R
\name{get_competitions}
\alias{get_competitions}
\title{Generate competition information}
\usage{
get_competitions(token)
}
\arguments{
\item{token}{API token generated by get_token}
}
\description{
get_competitons returns information from Impect's API including competition ID + name and season ID + name
}
