% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events.R
\name{get_events}
\alias{get_events}
\title{Generate match events from a match ID}
\usage{
get_events(token, matchID)
}
\arguments{
\item{token}{API token generated by get_token}

\item{matchid}{A match id}
}
\description{
get_events_from_matches returns event data from Impect's API for a single match
}
