% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events_from_comps.R
\name{get_events_from_comps}
\alias{get_events_from_comps}
\title{Generate match events from match IDs}
\usage{
get_events_from_comps(token, compids)
}
\arguments{
\item{token}{API token generated by get_token}

\item{compids}{A vector of compids}
}
\description{
get_events_from_comps returns event data from Impect's API from a vector of competition IDs. This is quicker for a single competition than using get_matches and then get_events_from_matches.
}
