% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events_from_matches.R
\name{get_events_from_matches}
\alias{get_events_from_matches}
\title{Generate match events from match IDs}
\usage{
get_events_from_matches(token, matchids)
}
\arguments{
\item{token}{API token generated by get_token}

\item{matchids}{A vector of matchids}
}
\description{
get_events_from_matches returns event data from Impect's API from a vector of match IDs
}
