% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_details.R
\name{get_match_details}
\alias{get_match_details}
\title{Generate match information}
\usage{
get_match_details(token, matchID)
}
\arguments{
\item{token}{API token generated by get_token}

\item{matchID}{Impect match ID}
}
\description{
get_match_details returns information about a single match details - including match_id, match_date, home + away teams and final score.
}
