% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matches.R
\name{get_matches}
\alias{get_matches}
\title{Generate match information}
\usage{
get_matches(token, comp)
}
\arguments{
\item{token}{API token generated by get_token}

\item{comp}{Impect competition ID}
}
\description{
get_matches returns information from Impect's API about match details - including match_id, match_date, home + away teams and final score.
}
