% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mins.R
\name{get_mins}
\alias{get_mins}
\title{Get minutes played by all players in a single matches}
\usage{
get_mins(token, matchID)
}
\arguments{
\item{token}{API token generated by get_token}

\item{matchID}{A single match ID}
}
\description{
get_mins returns a df with all minutes played in a single game
}
