% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mins_from_comps.R,
%   R/get_mins_from_matches.R
\name{get_mins_from_matches}
\alias{get_mins_from_matches}
\title{Get minutes played by all players in selected competitions}
\usage{
get_mins_from_matches(token, matchID)

get_mins_from_matches(token, matchID)
}
\arguments{
\item{token}{API token generated by get_token}

\item{compids}{A vector of competition IDs}

\item{matchids}{Vector of match IDs}
}
\description{
get_mins_from_matches returns a df with all minutes played in a vector of comp IDs, which can be generated with get_competition

get_mins_from_matches returns a df with all minutes played in a vector of match IDs, which can be generated with get_matches
}
